+++ mythtv/libs/libmyth/audio/spdifencoder.h
@@ -17,7 +17,7 @@ class MPUBLIC SPDIFEncoder
     SPDIFEncoder(const QString& muxer, AVCodecID codec_id);
     ~SPDIFEncoder();
     void WriteFrame(unsigned char *data, int size);
-    int  GetData(unsigned char *buffer, int &dest_size);
+    int  GetData(unsigned char *buffer, size_t &dest_size);
     int  GetProcessedSize() const { return m_size; };
     unsigned char *GetProcessedBuffer() { return m_buffer; };
     void Reset();
